Navigates to set edit screen from lift card
Removes the dialog for creating a max set and instead navigates to the edit set screen when a user taps on a lift card with no set.
This provides a consistent navigation pattern.

Also adds a test tag to the Leo onboarding card, and updates the empty lift card subtitle to be more descriptive.