fix[mobile]: fix crash when weight is 0f on linegraph
Simplifies the graph data structure by consolidating dot and line graph data into a single `GraphData` class.
This improves code readability and maintainability by reducing complexity in data handling for the graph.
Also fixes a bug where the bar graph height was not correctly computed. and was causing a crash if the weight/reps were 0f