Allows duplicating sets to today
Adds the ability to duplicate a set to the current day, overriding the default behavior of duplicating to the next available date.

This is achieved by introducing a `forceToday` flag in the `DuplicateSet` event, which, when true, ensures that the duplicated set is created with today's date.

This functionality is used when duplicating sets from the VariationItemCard in the workout screen.